added InitRoles functionality
